cambié el tipo de dato que regresa la función, por si llega a desbordarse el int.

+++ include/cli/get_total_archivos.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int get_total_archivos(const char* ruta_archivo_zip);
+long long int get_total_archivos(const char* ruta_archivo_zip);
